Revert "부산대_5조_순수웨딩_5주차_MASTER"

+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/image/ImageItemJPARepository.java
@@ -20,5 +20,5 @@ public interface ImageItemJPARepository extends JpaRepository<ImageItem, Long> {
     @Query("delete from ImageItem i where i.portfolio.id = :portfolioId")
     void deleteAllByPortfolioId(@Param("portfolioId") Long portfolioId);
 
-    void deleteAllByPortfolioPlannerId(Long id);
+    void deleteAllByPortfolioPlannerId(int id);
 }
